micuenta/requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.5
django-admin-tools==0.9.3
django-admin-tools-stats==0.9.0
django-bower==5.2.0
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.9
django-jazzmin==3.0.1
django-jsonfield==1.4.1
django-nvd3==0.10.1
django-qsstats-magic==1.1.0
djangorestframework==3.16.1
djcacheutils==3.0.0
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-memcached==1.62
python-nvd3==0.16.0
python-slugify==8.0.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
